- refactorized code - created selectnumberactivity view model and interfaces
