+++ app/src/main/java/com/example/randomuserapi/calls/randomuserusecase/RandomUserUseCaseRetrofit.kt
@@ -1,16 +1,21 @@
 package com.example.randomuserapi.calls.randomuserusecase
 
+import android.util.Log
+import com.example.randomuserapi.calls.usecaseclasses.randomuserentities.RandomUserEntity
 import com.example.randomuserapi.utils.ApiUrl
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
 class RandomUserUseCaseRetrofit {
 
-    val LOG_TAG = "RandomUserUseCaseRetrofit"
+    var logTag = this::class.java.toString()
 
     private var retrofit: Retrofit? = null
 
-    fun fetchRandomUserDataRetrofit(): RandomUserUseCaseInterface? {
+    fun retrofitBuilderRandomUserDataRetrofit(): RandomUserUseCaseInterface? {
 
         if (retrofit == null) {
             retrofit = Retrofit.Builder()
@@ -21,6 +26,25 @@ class RandomUserUseCaseRetrofit {
 
         return retrofit?.create(RandomUserUseCaseInterface::class.java)
 
+    }
+
+    fun fetchRandomUserDaraRetrofit(){
+        RandomUserUseCaseRetrofit().retrofitBuilderRandomUserDataRetrofit()?.getUserFromRandomUserApi()
+            ?.enqueue(object: Callback<RandomUserEntity> {
+                override fun onResponse(
+                    call: Call<RandomUserEntity>,
+                    response: Response<RandomUserEntity>
+                ) {
+                    if(response.isSuccessful){
+                        Log.d(logTag, "Obtained data using Retrofit from ${response.body().toString()}")
+                    }
+                }
+
+                override fun onFailure(call: Call<RandomUserEntity>, t: Throwable) {
+                    Log.d(logTag, t.message.toString())
+                }
+
+            })
     }
 
 }
