+++ app/src/main/java/com/example/randomuserapi/features/selecnumberofusers/selectnumberofusersviewmodel/SelectNumberOfUsersViewModelInterface.kt
@@ -1,4 +1,6 @@
 package com.example.randomuserapi.features.selecnumberofusers.selectnumberofusersviewmodel
 
 interface SelectNumberOfUsersViewModelInterface {
+    fun callRandomUserUseCase()
+    fun callRandomUserUseCaseRetrofit()
 }
